+++ app/helpers/bookingsync_portal/admin/application_helper.rb
@@ -1,6 +1,6 @@
 module BookingsyncPortal
   module Admin
-    module RentalsHelper
+    module ApplicationHelper
       def rental_details(rental)
         scope = 'bookingsync_portal.admin.rentals.rental'