+++ lib/generators/templates/initializers/bookingsync_portal.rb
@@ -10,7 +10,7 @@ BookingsyncPortal.setup do |config|
   config.message_bus_channel_scope = 'portal_app_name'
 
   # Allow to create a remote rental directly from this app.
-  config.create_remote_rental_from_app = false
+  config.create_remote_rental = false
 
   # customize account model class, can extend BookingsyncPortal::Account
   config.account_model = '::Account'
@@ -32,10 +32,6 @@ BookingsyncPortal.setup do |config|
 
   # customize rate model class, can extend BookingsyncPortal::Rate, set to nil if not used
   config.rate_model = '::Rate'
-
-  # handle creation of a remote rental on the remote account
-  def self.remote_rental_created(remote_rental)
-  end
 
   # handle synchronization of rentals after connection is made
   # def config.connection_created(connection)