+++ spec/dummy/app/models/remote_rental.rb
@@ -1,5 +1,2 @@
 class RemoteRental < BookingsyncPortal::RemoteRental
-  def name
-    uid
-  end
 end